requirements.txt
web3==6.0.0
textblob==0.17.1
click==8.1.3
python-dotenv==1.0.0